requirements.txt
pyairtable==3.1.1
tenacity==8.2.3
pytest-flask
pytest==7.4.0
pytest-cov==4.1.0
pydantic>=2.0
redis>=4.5.0
requests>=2.31.0
python-dotenv>=0.19.0
jsonschema==4.20.0
flask-cors==4.0.0
pytest-mock==3.14.0